LayoutTests/imported/w3c/csswg-test/css-shapes-1/shape-outside/values/shape-outside-inset-005-expected.txt


FAIL inset(10 20px 30px 40px) - invalid assert_equals: expected (object) null but got (string) ""
FAIL inset(10px 20 30px 40px) - invalid assert_equals: expected (object) null but got (string) ""
FAIL inset(10px, 20px 30 40px) - invalid assert_equals: expected (object) null but got (string) ""
FAIL inset(10px  20px 30px 40) - invalid assert_equals: expected (object) null but got (string) ""
FAIL inset(10 20 30 40) - invalid assert_equals: expected (object) null but got (string) ""
FAIL inset(10px, 20px, 30px, 40px) - invalid assert_equals: expected (object) null but got (string) ""
FAIL inset(10px, 20px, 30px 40px) - invalid assert_equals: expected (object) null but got (string) ""
FAIL inset(10px, 20px 30px 40px) - invalid assert_equals: expected (object) null but got (string) ""
FAIL inset(10px  20px, 30px, 40px) - invalid assert_equals: expected (object) null but got (string) ""
FAIL inset(10px  20px 30px, 40px) - invalid assert_equals: expected (object) null but got (string) ""
FAIL inset(foo bar blah woo) - invalid assert_equals: expected (object) null but got (string) ""
FAIL inset(10px roo 30px coo) - invalid assert_equals: expected (object) null but got (string) ""
FAIL inset(may 20px blah 40px) - invalid assert_equals: expected (object) null but got (string) ""
